update lupa-password features
