+++ components/form-lupa-password.module.css
@@ -4,13 +4,6 @@
   height: 156px;
   object-fit: cover;
 }
-.logoWrapper {
-  width: 333px;
-  display: none;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-}
 .masukkanEmail {
   flex: 1;
   position: relative;
@@ -28,38 +21,21 @@
   position: relative;
   line-height: 18px;
 }
-.contohJohndoegmailcom {
-  border: 0;
-  font-family: Poppins;
-  font-size: 14px;
-  background-color: transparent;
-  flex: 1;
-  position: relative;
-  line-height: 20px;
-  color: #8a8a8a;
-  text-align: left;
-}
-.fisearchIcon {
-  position: relative;
-  width: 24px;
-  height: 24px;
-  overflow: hidden;
-  flex-shrink: 0;
-}
 .form,
 .input {
   align-self: stretch;
   display: flex;
 }
 .input {
-  border-radius: 16px;
-  background-color: #fff;
   border: 1px solid #d0d0d0;
+  font-family: Poppins;
+  font-size: 14px;
+  background-color: #fff;
+  border-radius: 16px;
   flex-direction: row;
   padding: 12px 16px;
   align-items: center;
   justify-content: center;
-  gap: 16px;
 }
 .form {
   flex-direction: column;
@@ -95,14 +71,14 @@
   cursor: pointer;
   border: 0;
   padding: 12px 24px;
-  background-color: #7126b5;
+  background-color: #188e55;
   border-radius: 16px;
   flex-direction: row;
   justify-content: flex-start;
   gap: 4px;
 }
 .button:hover {
-  background-color: #4d2771;
+  background-color: #277029;
 }
 .buttonWrapper {
   padding: 8px 0 0;
@@ -111,11 +87,11 @@
 .buttonWrapper,
 .buttonWrapper:active,
 .buttonWrapper:hover,
-.frameParent,
+.forminput,
 .inner {
   flex-direction: column;
 }
-.frameParent,
+.forminput,
 .inner {
   display: flex;
   align-items: center;
@@ -127,7 +103,7 @@
   gap: 16px;
   font-size: 12px;
 }
-.frameParent {
+.forminput {
   flex: 1;
   overflow: hidden;
   gap: 32px;
@@ -141,20 +117,19 @@
   .buttonWrapper:hover {
     flex-direction: column;
   }
-  .frameParent {
+  .forminput {
     flex: unset;
     align-self: stretch;
   }
 }
 @media screen and (max-width: 960px) {
-  .logoIcon,
-  .logoWrapper {
+  .logoIcon {
     display: flex;
   }
   .buttonWrapper:hover {
     flex-direction: column;
   }
-  .frameParent {
+  .forminput {
     flex: unset;
     align-self: stretch;
   }
