+++ components/form-reset-password.js
@@ -2,15 +2,12 @@ import styles from "./form-reset-password.module.css";
 import { useState } from "react";
 import { useRouter } from "next/router";
 import axios from "axios";
-import { FaEye, FaEyeSlash } from "react-icons/fa";
 
 const FormResetPassword = () => {
   const [resetToken, setResetToken] = useState("");
   const [newPassword, setNewPassword] = useState("");
   const [verifyPassword, setVerifyPassword] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
-  const [showPassword, setShowPassword] = useState(false);
-  const [showVerifyPassword, setShowVerifyPassword] = useState(false);
 
   const router = useRouter();
 
@@ -24,14 +21,6 @@ const FormResetPassword = () => {
 
   const handleVerifyPasswordChange = (e) => {
     setVerifyPassword(e.target.value);
-  };
-
-  const toggleShowPassword = () => {
-    setShowPassword((prevShowPassword) => !prevShowPassword);
-  };
-
-  const toggleShowVerifyPassword = () => {
-    setShowVerifyPassword((prevShowVerifyPassword) => !prevShowVerifyPassword);
   };
 
   const handleSubmit = async (e) => {
@@ -60,75 +49,51 @@ const FormResetPassword = () => {
       setErrorMessage("An error occurred. Please try again later.");
     }
   };
+
   return (
-    <form className={styles.frameParent} onSubmit={handleSubmit}>
-      <div className={styles.logoWrapper}>
-        <img className={styles.logoIcon} alt="" src="/logo@2x.png" />
+    <form className={styles.inner} onSubmit={handleSubmit}>
+      <div className={styles.form}>
+        <div className={styles.token}>Token</div>
+        <input
+          className={styles.input}
+          id="resetToken"
+          value={resetToken}
+          onChange={handleResetTokenChange}
+          required
+          type="text"
+          placeholder="Masukkan Token"
+        />
       </div>
-      <div className={styles.framteksreset}>
-        <b className={styles.resetPassword}>Reset Password</b>
+      <div className={styles.form}>
+        <div className={styles.token}>Masukkan Password Baru</div>
+        <input
+          className={styles.input1}
+          placeholder="********"
+          type="password"
+          id="newPassword"
+          value={newPassword}
+          onChange={handleNewPasswordChange}
+          required
+        />
       </div>
-      <div className={styles.inner}>
-        <div className={styles.input}>
-          <div className={styles.masukkanToken}>Masukkan Token</div>
-          <input
-            className={styles.inputChild}
-            id="resetToken"
-            value={resetToken}
-            onChange={handleResetTokenChange}
-            required
-            type="text"
-            placeholder="Masukkan Token"
-          />
-        </div>
-        <div className={styles.input}>
-          <div className={styles.masukkanToken}>Masukkan Password Baru</div>
-          <div className={styles.masukkanPasswordBaruParent}>
-            <input
-              className={styles.masukkanPassword}
-              placeholder="Masukkan Password Baru"
-              type={showPassword ? "text" : "password"}
-              id="newPassword"
-              value={newPassword}
-              onChange={handleNewPasswordChange}
-              required
-            />
-            <button onClick={toggleShowPassword}>
-              {showPassword ? <FaEyeSlash /> : <FaEye />}
-            </button>
-          </div>
-        </div>
-        <div className={styles.input}>
-          <div className={styles.framteksreset}>
-            <div className={styles.ulangiPasswordBaru}>
-              Ulangi Password Baru
-            </div>
-          </div>
-          <div className={styles.masukkanPasswordParent}>
-            <input
-              className={styles.masukkanPassword}
-              type={showVerifyPassword ? "text" : "password"}
-              id="verifyPassword"
-              value={verifyPassword}
-              onChange={handleVerifyPasswordChange}
-              required
-              placeholder="Masukkan password"
-            />
-            <button onClick={toggleShowVerifyPassword}>
-              {showVerifyPassword ? <FaEyeSlash /> : <FaEye />}
-            </button>
-          </div>
-        </div>
-        <div className={styles.buttonWrapper}>
-          <button className={styles.button} type="submit">
-            <div className={styles.terbitkan}>Simpan</div>
-            <img className={styles.fiheartIcon} alt="" src="/fiheart.svg" />
-          </button>
-        </div>
+      <div className={styles.form}>
+        <div className={styles.token}>Ulangi Password Baru</div>
+        <input
+          className={styles.input1}
+          type="password"
+          placeholder="********"
+          id="verifyPassword"
+          value={verifyPassword}
+          onChange={handleVerifyPasswordChange}
+          required
+        />
       </div>
-      {errorMessage && <p>{errorMessage}</p>}
+      <div className={styles.buttonWrapper} type="submit">
+        <button className={styles.button}>
+          <div className={styles.terbitkan}>Simpan</div>
+        </button>
+      </div>
     </form>
-    
   );
 };
