+++ components/form-reset-password.module.css
@@ -1,185 +1,107 @@
-.logoIcon {
-    position: relative;
-    width: 266px;
-    height: 156px;
-    object-fit: cover;
-  }
-  .logoWrapper {
-    width: 333px;
-    display: none;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: flex-start;
-  }
-  .resetPassword {
-    flex: 1;
-    position: relative;
-    line-height: 36px;
-  }
-  .framteksreset {
-    align-self: stretch;
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-    justify-content: flex-start;
-  }
-  .masukkanToken {
-    align-self: stretch;
-    position: relative;
-    line-height: 18px;
-  }
-  .input,
-  .inputChild {
-    align-self: stretch;
-    display: flex;
-  }
-  .inputChild {
-    border: 1px solid #d0d0d0;
-    font-family: Poppins;
-    font-size: 14px;
-    background-color: #fff;
-    border-radius: 16px;
-    flex-direction: row;
-    padding: 12px 16px;
-    align-items: center;
-    justify-content: center;
-  }
-  .input {
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: flex-start;
-    gap: 4px;
-  }
-  
-  
-  .masukkanPasswordBaruParent {
-    align-self: stretch;
-    border-radius: 16px;
-    background-color: #fff;
-    border: 1px solid #d0d0d0;
-    display: flex;
-    flex-direction: row;
-    padding: 12px 16px;
-    align-items: center;
-    justify-content: center;
-    gap: 16px;
-  }
-  .masukkanPassword,
-  .ulangiPasswordBaru {
-    flex: 1;
-    position: relative;
-    line-height: 18px;
-  }
-  .masukkanPassword {
-    border: 0;
-    font-family: Poppins;
-    font-size: 14px;
-    background-color: transparent;
-    line-height: 20px;
-    color: #8a8a8a;
-    text-align: left;
-  }
-  .masukkanPasswordParent {
-    align-self: stretch;
-    border-radius: 16px;
-    background-color: #fff;
-    border: 1px solid #d0d0d0;
-    display: flex;
-    flex-direction: row;
-    padding: 12px 16px;
-    align-items: center;
-    justify-content: flex-start;
-    gap: 16px;
-  }
-  .terbitkan {
-    flex: 1;
-    position: relative;
-    font-size: 14px;
-    line-height: 20px;
-    font-weight: 500;
-    font-family: Poppins;
-    color: #fff;
-    text-align: center;
-  }
-  .fiheartIcon {
-    position: relative;
-    width: 16px;
-    height: 16px;
-    overflow: hidden;
-    flex-shrink: 0;
-    display: none;
-  }
-  .button,
-  .buttonWrapper {
-    align-self: stretch;
-    display: flex;
-    align-items: center;
-  }
-  .button {
-    cursor: pointer;
-    border: 0;
-    padding: 12px 24px;
-    background-color: #7126b5;
-    border-radius: 16px;
-    flex-direction: row;
-    justify-content: flex-start;
-    gap: 4px;
-  }
-  .button:hover {
-    background-color: #4d2771;
-  }
-  .buttonWrapper {
-    padding: 8px 0 0;
-    justify-content: center;
-  }
-  .buttonWrapper,
+.token {
+  align-self: stretch;
+  position: relative;
+  line-height: 18px;
+}
+.input {
+  border: 1px solid #d0d0d0;
+  font-family: Poppins;
+  font-size: 14px;
+  background-color: #fff;
+  border-radius: 16px;
+  flex-direction: row;
+  padding: 12px 16px;
+  align-items: center;
+  justify-content: center;
+}
+.form,
+.input,
+.input1 {
+  align-self: stretch;
+  display: flex;
+}
+.form {
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: flex-start;
+  gap: 4px;
+}
+.input1 {
+  border: 1px solid #d0d0d0;
+  background-color: #fff;
+  border-radius: 16px;
+  flex-direction: row;
+  padding: 12px 16px;
+  align-items: center;
+  justify-content: center;
+}
+.terbitkan {
+  flex: 1;
+  position: relative;
+  font-size: 14px;
+  line-height: 20px;
+  font-weight: 500;
+  font-family: Poppins;
+  color: #fff;
+  text-align: center;
+}
+.fiheartIcon {
+  position: relative;
+  width: 16px;
+  height: 16px;
+  overflow: hidden;
+  flex-shrink: 0;
+  display: none;
+}
+.button,
+.buttonWrapper {
+  align-self: stretch;
+  display: flex;
+  align-items: center;
+}
+.button {
+  cursor: pointer;
+  border: 0;
+  padding: 12px 24px;
+  background-color: #188e55;
+  border-radius: 16px;
+  flex-direction: row;
+  justify-content: flex-start;
+  gap: 4px;
+}
+.button:hover {
+  background-color: #277029;
+}
+.buttonWrapper {
+  padding: 8px 0 0;
+  justify-content: center;
+}
+.buttonWrapper,
+.buttonWrapper:active,
+.buttonWrapper:hover,
+.inner {
+  flex-direction: column;
+}
+.inner {
+  align-self: stretch;
+  display: flex;
+  padding: 0 45.9900016784668px 0 0;
+  align-items: center;
+  justify-content: center;
+  gap: 16px;
+  text-align: left;
+  font-size: 12px;
+  color: #000;
+  font-family: Poppins;
+}
+@media screen and (max-width: 1200px) {
   .buttonWrapper:active,
-  .buttonWrapper:hover,
-  .frameParent,
-  .inner {
+  .buttonWrapper:hover {
     flex-direction: column;
   }
-  .frameParent,
-  .inner {
-    display: flex;
-    align-items: center;
-    justify-content: center;
+}
+@media screen and (max-width: 960px) {
+  .buttonWrapper:hover {
+    flex-direction: column;
   }
-  .inner {
-    align-self: stretch;
-    padding: 0 45.9900016784668px 0 0;
-    gap: 16px;
-    font-size: 12px;
-  }
-  .frameParent {
-    flex: 1;
-    overflow: hidden;
-    gap: 32px;
-    text-align: left;
-    font-size: 24px;
-    color: #000;
-    font-family: Poppins;
-  }
-  @media screen and (max-width: 1200px) {
-    .buttonWrapper:active,
-    .buttonWrapper:hover {
-      flex-direction: column;
-    }
-    .frameParent {
-      flex: unset;
-      align-self: stretch;
-    }
-  }
-  @media screen and (max-width: 960px) {
-    .logoIcon,
-    .logoWrapper {
-      display: flex;
-    }
-    .buttonWrapper:hover {
-      flex-direction: column;
-    }
-    .frameParent {
-      flex: unset;
-      align-self: stretch;
-    }
-  }
+}
