+++ pages/lupa-password/index.js
@@ -1,10 +1,14 @@
-import FromLupaPassword from "../../components/form-lupa-password";
+import FromLupaPassword from "../../components/form-lupa-password"
 import styles from "./index.module.css";
 const LupaPassword = () => {
   return (
     <div className={styles.lupaPassword}>
-      <div className={styles.wLogin}>
-        <img className={styles.frame23301} alt="" src="/frame-2330-1@2x.png" />
+      <div className={styles.wLupaPassword}>
+        <img
+          className={styles.gambargradasiIcon}
+          alt=""
+          src="/gambargradasi@2x.png"
+        />
         <FromLupaPassword />
       </div>
     </div>
