--- pages/register/register-try.module.css
@@ -1,138 +1,0 @@
-.createAnAccount,
-.register {
-  align-self: stretch;
-  position: relative;
-  letter-spacing: 0.03em;
-}
-.register {
-  line-height: 52.51px;
-  font-weight: 900;
-}
-.createAnAccount {
-  font-size: 24px;
-  line-height: 26.25px;
-  color: #fffcf8;
-}
-.join {
-  align-self: stretch;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 28px;
-}
-.emailAddress {
-  border: 0;
-  font-family: Poppins;
-  font-size: 18px;
-  background-color: transparent;
-  flex: 1;
-  position: relative;
-  letter-spacing: 0.03em;
-  line-height: 26.25px;
-  color: #767676;
-  text-align: left;
-}
-.emailAddress:focus {
-  outline: 0;
-}
-.form {
-  border-radius: 7.5px;
-  background-color: #fff;
-  width: 419.77px;
-  height: 65px;
-  display: flex;
-  flex-direction: row;
-  padding: 8px 10px;
-  box-sizing: border-box;
-  align-items: center;
-  justify-content: flex-start;
-}
-.createAccount {
-  position: relative;
-  font-size: 24.75px;
-  letter-spacing: 0.03em;
-  line-height: 26.25px;
-  text-transform: capitalize;
-  font-weight: 500;
-  font-family: Poppins;
-  color: #fff;
-  text-align: left;
-}
-.button {
-  cursor: pointer;
-  border: 0;
-  padding: 11px 26px;
-  background-color: #000;
-  border-radius: 6px;
-  width: 423px;
-  display: flex;
-  flex-direction: row;
-  box-sizing: border-box;
-  align-items: center;
-  justify-content: center;
-}
-.communityInner,
-.registertry {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.communityInner {
-  align-self: stretch;
-  justify-content: flex-start;
-  gap: 18px;
-}
-.registertry {
-  position: relative;
-  background-color: #ddc9ff;
-  width: 100%;
-  padding: 111px 341.2900085449219px;
-  box-sizing: border-box;
-  justify-content: center;
-  text-align: center;
-  font-size: 41.25px;
-  color: #fff;
-  font-family: Poppins;
-}
-@media screen and (max-width: 1200px) {
-  .registertry {
-    padding-left: 200px;
-    padding-right: 200px;
-    box-sizing: border-box;
-  }
-}
-@media screen and (max-width: 960px) {
-  .register {
-    font-size: 35px;
-    line-height: 40px;
-  }
-  .createAnAccount,
-  .emailAddress {
-    font-size: 15px;
-  }
-  .form {
-    width: 100%;
-  }
-  .createAccount {
-    font-size: 20px;
-  }
-  .button {
-    width: 100%;
-  }
-  .registertry {
-    padding-left: 55px;
-    padding-right: 55px;
-    box-sizing: border-box;
-  }
-}
-@media screen and (max-width: 420px) {
-  .register {
-    font-size: 30px;
-    line-height: 30px;
-  }
-  .createAnAccount {
-    font-size: 13px;
-    line-height: 16px;
-  }
-}
