--- pages/reset-password/index.js
@@ -1,14 +1,0 @@
-import FormResetPassword from "../../components/form-reset-password";
-import styles from "./index.module.css";
-const ResetPassword = () => {
-  return (
-    <div className={styles.resetPassword}>
-      <div className={styles.wLogin}>
-        <img className={styles.frame23301} alt="" src="/frame-2330-1@2x.png" />
-        <FormResetPassword />
-      </div>
-    </div>
-  );
-};
-
-export default ResetPassword;
